# apps/commerce-api/openapi.json
{
  "components": {
    "schemas": {
      "AuthProvider": {
        "enum": [
          "password",
          "google",
          "apple"
        ],
        "title": "AuthProvider",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MerchantOnboardingRequest": {
        "properties": {
          "merchant_name": {
            "title": "Merchant Name",
            "type": "string"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "storefront_slug": {
            "title": "Storefront Slug",
            "type": "string"
          }
        },
        "required": [
          "merchant_name",
          "storefront_slug",
          "plan"
        ],
        "title": "MerchantOnboardingRequest",
        "type": "object"
      },
      "StaffInviteRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/TenantRole",
            "default": "staff"
          }
        },
        "required": [
          "email"
        ],
        "title": "StaffInviteRequest",
        "type": "object"
      },
      "StorefrontCreate": {
        "properties": {
          "slug": {
            "maxLength": 50,
            "minLength": 3,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug"
        ],
        "title": "StorefrontCreate",
        "type": "object"
      },
      "SwitchTenantRequest": {
        "properties": {
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id"
        ],
        "title": "SwitchTenantRequest",
        "type": "object"
      },
      "TenantRole": {
        "enum": [
          "owner",
          "manager",
          "staff"
        ],
        "title": "TenantRole",
        "type": "string"
      },
      "TenantStatus": {
        "enum": [
          "provisioning",
          "onboarding",
          "active",
          "suspended",
          "offboarding",
          "archived"
        ],
        "title": "TenantStatus",
        "type": "string"
      },
      "TenantStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TenantStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "TenantStatusUpdate",
        "type": "object"
      },
      "UserCreate": {
        "additionalProperties": false,
        "properties": {
          "auth_provider": {
            "$ref": "#/components/schemas/AuthProvider",
            "default": "password"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "required": [
          "email"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CruxNexus Commerce API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/identity/login": {
      "post": {
        "description": "Authenticate user and return a global merchant access token.",
        "operationId": "login_api_v1_identity_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Login Api V1 Identity Login Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/identity/storefronts": {
      "get": {
        "operationId": "list_accessible_storefronts_api_v1_identity_storefronts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Accessible Storefronts",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/identity/switch-tenant": {
      "post": {
        "description": "Switch tenant context with full membership verification.\n\nThis endpoint verifies:\n1. User has active membership to the requested tenant\n2. Tenant is not archived\n3. Membership has a valid role\n\nOnly then does it issue a tenant-scoped token.",
        "operationId": "switch_tenant_api_v1_identity_switch_tenant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Switch Tenant",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/identity/users": {
      "post": {
        "description": "Register a global user account.",
        "operationId": "create_user_api_v1_identity_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/identity/users/me": {
      "get": {
        "operationId": "get_current_user_api_v1_identity_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Current User",
        "tags": [
          "identity"
        ]
      }
    },
    "/api/v1/merchant/storefronts": {
      "get": {
        "operationId": "merchant_list_storefronts_api_v1_merchant_storefronts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merchant List Storefronts",
        "tags": [
          "merchant"
        ]
      },
      "post": {
        "operationId": "merchant_create_storefront_api_v1_merchant_storefronts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorefrontCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merchant Create Storefront",
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/v1/merchant/storefronts/{tenant_id}": {
      "get": {
        "operationId": "merchant_get_storefront_api_v1_merchant_storefronts__tenant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merchant Get Storefront",
        "tags": [
          "merchant"
        ]
      }
    },
    "/api/v1/onboarding/merchant": {
      "post": {
        "description": "Complete merchant onboarding in one atomic transaction.\n\nCreates:\n- Merchant Account\n- Merchant Account User relationship\n- Subscription\n- Tenant (Storefront)\n- Merchant Account Tenant ownership\n- Tenant Membership (OWNER)\n\nThis is the authoritative merchant bootstrap operation.",
        "operationId": "onboard_merchant_api_v1_onboarding_merchant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantOnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Onboard Merchant",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/storefronts": {
      "post": {
        "operationId": "create_storefront_api_v1_storefronts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorefrontCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Storefront",
        "tags": [
          "storefronts"
        ]
      }
    },
    "/api/v1/storefronts/{tenant_id}": {
      "get": {
        "operationId": "get_storefront_api_v1_storefronts__tenant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Storefront",
        "tags": [
          "storefronts"
        ]
      }
    },
    "/api/v1/tenants/{tenant_id}/members": {
      "post": {
        "operationId": "invite_staff_member_api_v1_tenants__tenant_id__members_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invite Staff Member",
        "tags": [
          "storefronts"
        ]
      }
    },
    "/api/v1/tenants/{tenant_id}/status": {
      "patch": {
        "operationId": "update_tenant_status_api_v1_tenants__tenant_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Tenant Status",
        "tags": [
          "storefronts"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Check"
      }
    },
    "/health/live": {
      "get": {
        "operationId": "liveness_check_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Liveness Check"
      }
    },
    "/health/ready": {
      "get": {
        "description": "Readiness check - verifies DB and Redis connectivity.",
        "operationId": "readiness_check_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness Check"
      }
    }
  }
}